// src/component/page/bruker-detaljer/deltaker-detaljer/types.ts
export enum EndringType {
  LEGG_TIL_OPPSTARTSDATO = 'LEGG_TIL_OPPSTARTSDATO',
  ENDRE_OPPSTARTSDATO = 'ENDRE_OPPSTARTSDATO',
  FORLENG_DELTAKELSE = 'FORLENG_DELTAKELSE',
  AVSLUTT_DELTAKELSE = 'AVSLUTT_DELTAKELSE',
  DELTAKER_IKKE_AKTUELL = 'DELTAKER_IKKE_AKTUELL',
  ENDRE_DELTAKELSE_PROSENT = 'ENDRE_DELTAKELSE_PROSENT',
  ENDRE_SLUTTDATO = 'ENDRE_SLUTTDATO',
  ENDRE_SLUTTAARSAK = 'ENDRE_SLUTTAARSAK'
}